Update C2_ET_SQL_CREATE_v0.2.sql
combio tipo per i CFUPrevisti

+++ DB/C2_ET_SQL_CREATE_v0.2.sql
@@ -38,7 +38,7 @@ on delete cascade
 CREATE TABLE TIROCINIO (
 CODTIROCINIO INT auto_increment,
 DATAINIZIOTIROCINO date not null,
-CFUPREVISTI tinyint(3) not null,
+CFUPREVISTI int(3) not null,
 COMPETENZE longtext not null,
 COMPETENZEACQUISIRE longtext not null,
 ATTIVITAPREVISTE longtext not null,
